+++ src/main/java/com/pedrofranceschi/orderapi/repository/CidadeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CidadeRepository extends JpaRepository<Cidade, Integer>{
+public interface CidadeRepository extends JpaRepository<Cidade, Long>{
 
-    List<Cidade> findByEstadoId(Integer estadoId);
+    List<Cidade> findByEstadoId(Long estadoId);
 }
