+++ src/main/java/com/pedrofranceschi/orderapi/repository/ClienteRepository.java
@@ -3,7 +3,7 @@ package com.pedrofranceschi.orderapi.repository;
 import com.pedrofranceschi.orderapi.entities.Cliente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClienteRepository extends JpaRepository<Cliente, Integer>{
+public interface ClienteRepository extends JpaRepository<Cliente, Long>{
 
     Cliente findByNome(String nome);
 }
