+++ src/main/java/com/pedrofranceschi/orderapi/repository/EstadoRepository.java
@@ -4,7 +4,7 @@ import com.pedrofranceschi.orderapi.entities.Cliente;
 import com.pedrofranceschi.orderapi.entities.Estado;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstadoRepository extends JpaRepository<Estado, Integer> {
+public interface EstadoRepository extends JpaRepository<Estado, Long> {
 
     Estado findByUf(String uf);
 }
